requirements.txt
streamlit
pandas
plotly
supabase
Pillow